backend/requirements.txt
fastapi
uvicorn
pyyaml
requests
pyotp
cryptography
loguru
httpx
